+++ AndroidTimetableView/TimetableView/res/layout/item_timetable.xml
@@ -9,8 +9,8 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@drawable/item_corner_style"
-        android:gravity="center"
-        android:padding="3dp"
+        android:gravity="center_horizontal|top"
+        android:padding="2dp"
         android:textColor="@color/app_white"
         android:textSize="12sp" />
