+++ AndroidTimetableView/TimetableView/res/layout/timetable_layout.xml
@@ -11,7 +11,6 @@
         android:id="@+id/id_datelayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@color/app_course_bg_gray"
         android:baselineAligned="false"
         android:gravity="center_vertical"
         android:orientation="horizontal">
