+++ AndroidTimetableView/TimetableView/res/values/dimens.xml
@@ -4,8 +4,8 @@
 <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
 	<dimen name="weekItemHeight">55dp</dimen>  
-    <dimen name="weekItemMarTop">2dp</dimen>
-    <dimen name="weekItemMarLeft">2dp</dimen>
+    <dimen name="weekItemMarTop">3dp</dimen>
+    <dimen name="weekItemMarLeft">3dp</dimen>
     <dimen name="headHeight">50dp</dimen>
     <dimen name="shadow_width">2dp</dimen>
 </resources>
