+++ AndroidTimetableView/TimetableView/src/com/zhuangfei/timetable/listener/OnItemBuildAdapter.java
@@ -1,6 +1,7 @@
 package com.zhuangfei.timetable.listener;
 
 import android.graphics.drawable.GradientDrawable;
+import android.text.TextUtils;
 import android.widget.FrameLayout;
 import android.widget.TextView;
 
@@ -13,11 +14,13 @@ import com.zhuangfei.timetable.model.Schedule;
 public class OnItemBuildAdapter implements ISchedule.OnItemBuildListener {
     @Override
     public String getItemText(Schedule schedule, boolean isThisWeek) {
-        if(isThisWeek){
-            return schedule.getName()+"@"+schedule.getRoom();
+        if(schedule==null||schedule.getName()==null||schedule.getRoom()==null) return "default";
+        String r=schedule.getName()+"@"+schedule.getRoom();
+        if(!isThisWeek){
+            r="[非本周]"+r;
         }
 
-        return schedule.getName();
+        return r;
     }
 
     @Override
@@ -27,6 +30,5 @@ public class OnItemBuildAdapter implements ISchedule.OnItemBuildListener {
 
     @Override
     public void onItemUpdate(FrameLayout layout, TextView textView, TextView countTextView, Schedule schedule, GradientDrawable gd) {
-
     }
 }
