--- AndroidTimetableView/TimetableView/src/com/zhuangfei/timetable/model/ScheduleTheme.java
@@ -1,8 +1,0 @@
-package com.zhuangfei.timetable.model;
-
-/**
- * Created by Liu ZhuangFei on 2018/7/28.
- */
-
-public class ScheduleTheme {
-}
