+++ AndroidTimetableView/app/src/main/AndroidManifest.xml
@@ -18,17 +18,6 @@
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
-        <activity android:name=".SimpleActivity" />
-        <activity android:name=".ItemStyleActivity" />
-        <activity android:name=".AttrActivity" />
-        <activity android:name=".ElasticActivity" />
-        <activity android:name=".MainActivity">
-
-        </activity>
-        <activity android:name=".SlideActivity" />
-        <activity android:name=".DateActivity" />
-        <activity android:name=".ColorPoolActivity" />
-        <activity android:name=".NonViewActivity" />
         <activity android:name=".NewFeatureActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
