+++ AndroidTimetableView/app/src/main/res/drawable-hdpi/bg.jpg
